+++ src/komm/__init__.py
@@ -61,6 +61,9 @@ from ._error_control_convolutional.ViterbiStreamDecoder import ViterbiStreamDeco
 from ._error_control_decoders.BCJRDecoder import BCJRDecoder
 from ._error_control_decoders.BerlekampDecoder import BerlekampDecoder
 from ._error_control_decoders.ExhaustiveSearchDecoder import ExhaustiveSearchDecoder
+from ._error_control_decoders.GaussianEliminationDecoder import (
+    GaussianEliminationDecoder,
+)
 from ._error_control_decoders.ReedDecoder import ReedDecoder
 from ._error_control_decoders.SCDecoder import SCDecoder
 from ._error_control_decoders.SyndromeTableDecoder import SyndromeTableDecoder
@@ -189,6 +192,7 @@ __all__ = [
     "BCJRDecoder",
     "BerlekampDecoder",
     "ExhaustiveSearchDecoder",
+    "GaussianEliminationDecoder",
     "ReedDecoder",
     "SCDecoder",
     "SyndromeTableDecoder",