+++ tests/error_control_decoders/test__decoders.py
@@ -16,6 +16,7 @@ spc = komm.SingleParityCheckCode(3)
     [
         [block, komm.SyndromeTableDecoder],
         [block, komm.ExhaustiveSearchDecoder],
+        [block, komm.GaussianEliminationDecoder],
         [terminated, komm.ViterbiDecoder],
         [terminated, komm.BCJRDecoder],
         [bch, komm.BerlekampDecoder],